Add two private attribute _ssh and _sftp
Multiple ssh and sftp are opened and closed during using the SSHClient module can brought several problems: (1) increasing connection load and more sensitive to the stability of the connection. (2) sessions, for some reason are not synchronized immediately. (3) repetition of creating and closing SSH and SFTP blocks in most of the methods. Adding these two private attributes is more desirable.